requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
drf-yasg>=1.21
psycopg2-binary>=2.9   # optional, only if using Postgres
